refactor contract artifacts into their own folders
