+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/commands/use-cases/create-user/contracts/create-user.command.ts
@@ -1,0 +1,6 @@
+import { ICommand } from "../../../../../../libs/ddd/command.interface";
+import { UserEntity } from "../../../../entities/user.entity";
+
+export interface CreateUserCommand
+  extends ICommand,
+    Pick<UserEntity, "email" | "username" | "firstName" | "lastName"> {}
