+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/commands/use-cases/create-user/contracts/create-user.controller.interface.ts
@@ -1,5 +1,5 @@
-import { FailResponseDto } from "../../../../../libs/api/fail.response.dto";
-import { IController } from "../../../../../libs/ddd/controller.interface";
+import { FailResponseDto } from "../../../../../../libs/api/fail.response.dto";
+import { IController } from "../../../../../../libs/ddd/controller.interface";
 import { CreateUserRequestDto } from "./create-user.request-dto";
 import { CreateUserResponseDto } from "./create-user.response-dto";
