+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/commands/use-cases/create-user/contracts/create-user.request-dto.ts
@@ -1,6 +1,6 @@
 import { IsEmail, Length, MinLength } from "class-validator";
-import { UserEntity } from "../../../entities/user.entity";
-import { IRequestDto } from "../../../../../libs/api/request.dto.interface";
+import { IRequestDto } from "../../../../../../libs/api/request.dto.interface";
+import { UserEntity } from "../../../../entities/user.entity";
 
 export class CreateUserRequestDto
   implements
