+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/commands/use-cases/create-user/contracts/create-user.response-dto.ts
@@ -1,5 +1,5 @@
-import { OkResponseDto } from "../../../../../libs/api/ok.response.dto";
-import { UserEntity } from "../../../entities/user.entity";
+import { OkResponseDto } from "../../../../../../libs/api/ok.response.dto";
+import { UserEntity } from "../../../../entities/user.entity";
 
 interface DtoUserProps
   extends Pick<
