+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/commands/use-cases/create-user/contracts/create-user.use-case.interface.ts
@@ -1,0 +1,9 @@
+import { IUseCase } from "../../../../../../libs/ddd/use-case.interface";
+import { UserEntity } from "../../../../entities/user.entity";
+import { UserRepositoryError } from "../../../../repository/contracts/user.repository.interface";
+import { CreateUserCommand } from "./create-user.command";
+
+export type CreateUserUseCaseError = UserRepositoryError;
+
+export interface ICreateUserUseCase
+  extends IUseCase<CreateUserCommand, UserEntity, CreateUserUseCaseError> {}
