--- 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/commands/use-cases/create-user/create-user.use-case.interface.ts
@@ -1,9 +1,0 @@
-import { IUseCase } from "../../../../../libs/ddd/use-case.interface";
-import { UserEntity } from "../../../entities/user.entity";
-import { UserRepositoryError } from "../../../repository/user.repository.interface";
-import { CreateUserCommand } from "./create-user.command";
-
-export type CreateUserUseCaseError = UserRepositoryError;
-
-export interface ICreateUserUseCase
-  extends IUseCase<CreateUserCommand, UserEntity, CreateUserUseCaseError> {}
