+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/repository/contracts/user.repository.interface.ts
@@ -1,5 +1,5 @@
-import { Result } from "../../../libs/utils/result";
-import { UserEntity } from "../entities/user.entity";
+import { Result } from "../../../../libs/utils/result";
+import { UserEntity } from "../../entities/user.entity";
 
 export type UserRepositoryError =
   | "EMAIL_ALREADY_EXISTS"
