+++ 0_From_Code_First_To_First_Principles/1_code_first/assignment/src/modules/user/repository/user.repository.ts
@@ -3,7 +3,7 @@ import { UserEntity } from "../entities/user.entity";
 import {
   IUserRepository,
   UserRepositoryError,
-} from "./user.repository.interface";
+} from "./contracts/user.repository.interface";
 import { MikroORM, PostgreSqlDriver } from "@mikro-orm/postgresql";
 import { errorToString } from "../../../libs/utils/error-to-string";
